ACTION-3580: Remove <use> element instance tree.
As part of this, add references to Shadow DOM, doing
a part of ACTION-3413. This needs further review and
changes for the style parts, but it's a start. We can
probably remove or make some parts of the spec text
about instance trees informative, and refer to
Shadow DOM instead.