[linking] Addressing issues 6,7 and 8
Used the opportunity to clean the paragraph SVG fragment identifiers to
introduce a bit more structure and break the wall of text.

Issue 6 - Referenced CSS Transforms spec and as discussed on Paris 2015
F2F meeting spec'd that svgView parameters apply in definition sequence

Issue 7 - As discussed and agreed on Paris 2015 F2F meeting at this
point we don't want to bring addition transform parameters and styles
(such as transform-origin, perspective, preserve-3d). Based on the
further feedback we'll consider those changes in the future.

Issue 8 - Reflected actual browser support. Spaces are supported in all
browsers (and we couldn't identify obvious reason they shouldn't) and
semicolon escaping in not supported in WebKit/Blink but we've discussed
that URL fragment processing should adhere to general URL processing
rules and referenced escaping rules.