[linking] Address issue 9
As discussed on Paris 2015 F2F meeting we've agreed to simplify this and
apply svgView to the root svg element and not closest svg ancestor.
While there might be some niche scenarios this might have enabled, the
complexity and unpredictability this introduces for the authors is not
desired. In the future we can revisit this based on further feedback.
The current specification should fully address existing
(non-theoretical) scenarios.