[linking] Deprecating viewTarget attribute on the <view> element
RESOLUTION: We will remove the <view viewTarget> attribute, its DOM
property, and the corresponding view spec argument.

A bit of context here - on the meeting we've discussed that the
attribute was create in hope to allow some sort of automatic
positioning, sizing and clipping of the views based on viewTarget
elements. However we've confirmed that no browser engine (Blink, Gecko,
Edge) implements it this way and viewTarget has no functional effect on
the presentation of SVG content. We've also confirmed there are no open
issues on Blink, Gecko, Edge and no indication of author interest
(Stackoverflow, GitHub). Even with good intentions the attribute wasn't
spec'd enough in SVG 1.1 or SVG 2 and we've agreed to remove the
attribute (+ DOM, svgView specification).