Cross-references and external resources:
Define common behavior for processing all URLs, with reference to HTML's potential-CORS fetch algorithm.  Some normative details which were previously unspecified:

- Consistent with the Masking spec, references from style properties should support anonymous CORS.
-  All subresources with specific element targets must be processed in secure static mode.
- Embedded images must be processed in secure static or animated mode.

Also clean-up various other out-of-date bits in the Linking chapter to match changes to other chapters:

- textPath may reference basic shapes
- hatch accepts an href attribute
- xml:base is only valid in XML documents, and HTML `<base>` is valid in SVG.

Add an example to the `image` section demonstrating how URLs with target fragment refer to the entire document, even when that document is the same one with the image in it.

Closes #90 (Update href attribute descriptions to reference HTML5.1) and addresses some of the discussion in pull request #206.