Cursors: mark <cursor> element deprecated, update other prose
This is consistent with CSS fully supporting custom cursors, and with HTML specifically disallowing cursor references to cursor elements embedded in an HTML file.

Also add clarifying text to the Linking chapter on how to process URLs when the valid target can be either a specific element or an image file (i.e., cursor and mask properties).

Closes #174