Text chapter: Review & edits
Including some final edits with respect to issues #200 and #210, as well as general typo fixes, formatting, and clarification.

Changes the definition and example of `shape-margin` to reflect the fact that this is a property of the excluded element, not the element from which the exclusion is subtracted.  (attn @Tavmjong)

Adds a note about ::before and ::after being possible in the future (see #125).

Makes ::selection and :focus styles a requirement for interactive user agents, while leaving open how that is implemented.  (Most web browsers use background color as part of a selection style, and we don't have a way of specifying background color on a span of text!)

Not the full edit I'd like to have done, but hopefully caught most normative issues.