Copy Editing: Eradicate the usage of informal `you'; clean up surrounding text
(Ultimately, this commit should not change the meaning of the text.)

Because this is a standard, it is probably best not to employ the rather
informal second person, `you'. This commit removes that usage without
descending into overtly bureaucratic language.

Furthermore, as a matter of practicality or even necessity, some of
the nearby text has also been improved (according to this author);
this includes rewording, re-styling, the removal of typos, and the
fixing of grammar.