Update stylesheets for local dev
- update the included copy of base.css to match
  https://www.w3.org/StyleSheets/TR/2016/base.css
- Replace the stylesheets in the head of each of the SVG 2 chapter files
  with the normal Editor's Draft styles
 (local copy of W3C ED/base + svg.css)
- Update the comment in the head to explain
  what will happen in processing.
- Change the processing script
  so that it looks for a `data-keep` attribute
  to decide whether to leave a stylesheet link,
  instead of only looking for print stylesheets.
- Change the publish script to also keep the `data-keep` stylesheets.

These changes mean that it is easier to review changes to SVG 2
without having to run a complete local build process.

It also makes it easier in future to add style overrides
as `<link>` instead of `<style>`.

The changes shouldn't have any effect on
the current processed versions of the specs.