Reintroduce segment-completing closepath as a theoretical concept
Because we use it to define the equivalent paths of basic shapes.

I've changed the terminology to refer to the segment-completing close path
as an _operation_ instead of a _command_,
and have added a short sub-section defining it.

I also included a very short note about
the proposal for adding it to the path data syntax,
in a future version of the specification.