Add stub files for removed chapters/appendices
So that links to URLs of the form
https://www.w3.org/TR/SVG/backward.html won't 404.

Where relevant, the stubs include a very brief description
of what happened to the content,
with links to its current location.
There are also links back to the SVG11 TR version of the chapter,
and to the main SVG2 Overview page.

To include these files in the final build,
without messing up building the actual spec,
they are included as "resource" files in publish.xml.
This means that they are copied as-is to the final directory,
with no build system modifications.