No value is illegal
But many values are <a>invalid</a> and must be <a>ignored</a>!

Changes the prose describing non-negative value constraints
to use the CSS 3 terms.

"invalid" is linked to the existing definition of "invalid value"
in the types chapter.

"ignored" links directly to the CSS Syntax definition
(ED version for now; the published CR is out of date.)

Closes #639 and tidies up changes made in #637.