Update painting.html stroke-width
(1)  Specify that the stroke-width is centered on the path, rather than inside or outside the shape.
(2)  Correct(?) the stroke-width of a dash.

(A)  It is expected that the location of the stroke's own paint will be controllable in the next SVG edition, but the default/initial position will still need to be defined.

(B)  The algorithm for painting a dash collects all points on a line perpendicular to the path that are within a distance of stroke-width.  Because the path has two sides, this effectively doubles the stroke-width.  Either that should be corrected (as I proposed here) or it should be made explicit that dashes are double-thick.