Automate publication of Editor's Drafts to GitHub Pages
This adds a GitHub workflow to automate the publication of the Editor's Drafts
in the repository to GitHub Pages. This is the first step towards transferring
the `svgwg.org` domain name to the W3C, as envisioned in:
https://github.com/w3c/svgwg/issues/1006#issuecomment-3293231900

Under the hoods, the workflow runs `make`, then a new "quick and dirty" bash
script that copies the results of the build to a `deploy` folder with the
appropriate structure, and then actually copies the `deploy` folder to the
`gh-pages` branch for publication.

The workflow also runs on pull requests as a way to test that the update would
not break the build. It does *not* deploy the result when it runs on a PR.