Bikeshed Conversion #1036
The first pass is done. Yoohoo!

BIKESHED:
Bikeshed Documentation
https://speced.github.io/bikeshed/

Command for compiling the spec:
    bikeshed spec Overview.bs

You need a local install as it is composed of multiple
files.

SPEC ORGANIZATION:
Overview.bs is linking to markdown version of the
different chapters. The HTML version files are kept
in the repo for now to make sure, information has
not been lost until the confidence is high enough
that we can remove them.

In the future, Tab Atkins will make it possible to
generate a multipage version of the spec.

All chapters follow the previous organization
except toward the appendices.

WORK TO BE DONE:
Some questions and TODO:

* TODO: Fix references. Successfully generated, with 3389 linking errors
* TODO: Some of the summary for elements were done automagically. We need to double check this. Search @@ in the prose.
* TODO: Check that all images are correctly integrated
* TODO: Check that all markup examples are correctly inserted
* TODO: Rewrite step by step the paragraph so that they are more bikeshedy
* QUESTION: Bikeshed handles a lot of the references building, maybe we can get rid of them in the prose as they will be automatically added. See TODO below.
* TODO: Check if the IDL references is similar to the previous one
* TODO: Check if the element index is correct
* TODO: Check if the attribute index is correct
* TODO: Check if the property index is correct
* TODO: Check if the references index is correct